hubble: move container/priority_queue to hubble/relay/queue
This priority queue implementation is not yet used and only relevant for
hubble-relay.
