+++ pkg/hubble/relay/queue/priority_queue.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package container
+package queue
 
 import (
 	"container/heap"
