+++ pkg/hubble/relay/queue/priority_queue_test.go
@@ -14,7 +14,7 @@
 
 // +build !privileged_tests
 
-package container
+package queue
 
 import (
 	"testing"
